Removing TimeSparseArray
The only added value in this subclass comes from the added methods to
binary search on its keys. Pulling these methods up to the base class
LongSparseArray so they can be useful for all its instances.

Renamed variables and methods to be more generic and replaced the
handwritten binary search with an equivalent call to
Arrays.binarySearch to reduce maintenance overhead.

All occurrences of TimeSparseArray have been replaced with
LongSparseArray. No behavior change is expected.

Test: Builds, boots.
Test: atest FrameworksCoreTests:LongSparseArrayTest

Bug: 282264041
Change-Id: I6e87560ad5afe85118c9ffaac9ea8a506a181af0