Initialize hearing aid info if the device matches hearing aid scan filters
We only show connected hearing devices in the hearing devices page now.
When user pairing a device from pairing page and back to the hearing
devices page after the device is bonded, it's confusing no device shown
in the list because we can't get hearing aid info and thus don't know
it's a hearing aid while the profiles are not connected yet.

Init the device with empty hearing aid info if it matches the scan
filter. By doing this we can know it's a hearing aid before hearing aid
related profiles are connected and can show the connecting hearing aid
devices in the a11y hearing devices page to avoid confusion.

Bug: 283268686
Test: atest CachedBluetoothDeviceTest
Test: atest CachedBluetoothDeviceManagerTest
Test: atest HearingAidDeviceMangerTest
Change-Id: I3d3916f85a90c3f610a1d59ae244126788de0b2d