Remove vkDeviceWaitIdle call when destroying VulkanSurface
vkDeviceWaitIdle requires external synchronization of all queues created from the VkDevice, which would mean that we would need to add locking to the AHB uploader queue for this to be thread-safe.

However, we should not need to wait for queue submissions for the AHB uploader at all for destroying the output surface, because the AHB uploader queue is meant for swizzling bitmaps - there's no dependency on the output surface.

So, it is just sufficient to wait on submission of the graphics queue, which is currently thread-safe as all graphics queue operations occur only on renderthread.

Bug: 288201932
Test: builds
Test: com.android.launcher3.ui.TaplTestsLauncher3#testDragShortcutToWorkspaceCell 99 times
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:727351503200b9b983e314aa7e3fea880437561c)
Merged-In: Ia3ca1600f22941d9900b2e5451f7572a79abf049
Change-Id: Ia3ca1600f22941d9900b2e5451f7572a79abf049