Reconnect if weaver service connection dies
For robustness reasons, reconnect to the Weaver service if it dies.
This should not be relied on, but it's easy to do and is the same thing
that LockSettingsService does for the Gatekeeper service.

Bug: 287705522
Test: Treehugger
Change-Id: Ic06c45353c6b7282e2f22eef13ea1c6e36fe5c58
Merged-In: Ic06c45353c6b7282e2f22eef13ea1c6e36fe5c58
(cherry picked from commit 66cc8241fbfd1ae08822e7f975e5f14a304ab345)